add ppc disassembly to jit results viewer windowTurn the JIT into a class. Indentation cleanup will be in a separate CL, and the block cache will later be split out into its own class. Cannot detect any speed difference whatsoever - the games spend 99.9% of their time in already jitted code anyway. I have a good reason for doing this - see upcoming changes.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@1585 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/PowerPC/PPCAnalyst.cpp
@@ -364,7 +364,7 @@ bool Flatten(u32 address, int *realsize, BlockStats *st, BlockRegStats *gpa, Blo
 	}
 
 	gpa->any = true;
-	for (size_t i = 0; i < num_inst; i++)
+	for (int i = 0; i < num_inst; i++)
 	{
 		UGeckoInstruction inst = code[i].inst;
 		if (PPCTables::UsesFPU(inst))
